Abstract and refactor tests and request
