+++ app/Http/Requests/AdminCentreUserRequest.php
@@ -6,7 +6,7 @@ use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 use Request;
 
-class AdminNewCentreUserRequest extends FormRequest
+class AdminCentreUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
